adding vcpkg support

+++ src/projname/projname_main.cpp
@@ -6,5 +6,5 @@
 int main(int argc, char const* argv[]) {
     std::set_new_handler([] { std::terminate(); });
 
-    return projname::run({argv, argc});
+    return projname::run({argv, static_cast<std::size_t>(argc)});
 }
